Move setTownFlags to town util
