+++ src/main/java/com/gmail/goosius/siegewar/utils/SiegeWarTownUtil.java
@@ -1,5 +1,6 @@
 package com.gmail.goosius.siegewar.utils;
 
+import com.gmail.goosius.siegewar.settings.SiegeWarSettings;
 import com.palmergames.bukkit.towny.object.Town;
 import com.palmergames.bukkit.towny.object.TownBlock;
 import com.palmergames.bukkit.towny.object.TownBlockType;
@@ -8,7 +9,7 @@ import com.palmergames.bukkit.towny.object.TownyPermission.PermLevel;
 import com.palmergames.bukkit.towny.object.TownyPermissionChange.Action;
 
 /**
- * Util class containing useful methods for town permissions etc.
+ * Util class containing methods related to town flags/permssions.
  */
 public class SiegeWarTownUtil {
     public static void disableTownPVP(Town town) {
@@ -52,4 +53,17 @@ public class SiegeWarTownUtil {
         }
         town.save();
     }
+
+    /**
+	 * Sets pvp and explosions in a town to the desired setting, if enabled in the config.
+	 * 
+	 * @param town The town to set the flags for.
+	 * @param desiredSetting The value to set pvp and explosions to.
+	 */
+	public static void setTownFlags(Town town, boolean desiredSetting) {
+		if (town.getPermissions().pvp != desiredSetting && SiegeWarSettings.getWarSiegePvpAlwaysOnInBesiegedTowns())
+			town.getPermissions().pvp = desiredSetting;
+		if (town.getPermissions().explosion != desiredSetting && SiegeWarSettings.getWarSiegeExplosionsAlwaysOnInBesiegedTowns())
+			town.getPermissions().explosion = desiredSetting;
+	}
 }
